Combine the publishing-api test helpers
Makes the test helpers backwards compatible to reduce the risk of this
being a breaking change.
